Moved popup.js into Popup.jsx
