+++ package.json
@@ -13,6 +13,7 @@
     "babel-preset-es2015": "6.18.0",
     "babel-preset-react": "6.16.0",
     "babel-runtime": "^6.20.0",
+    "copy-webpack-plugin": "^4.0.1",
     "css-loader": "0.26.1",
     "extract-text-webpack-plugin": "^1.0.1",
     "file-loader": "0.9.0",
